add apy to pair page
this would be useful on the leaderboard to be able to sort by, but this is
pretty useful for eye balling anyway.

it's an incredibly clumsy calculation, really, but it's the same thing uniswap
does to make returns look insane. it annualizes returns over 24h based on
current liquidity, but it's something anyway. just having 'fees' listed for
users means they end up doing something like this on their own to decide which
pairs to invest in, so this is more useful even if it's imperfect, fees on its
own isn't a very meaningful number, only in relation to the liquidity can
returns be calculated.